csimanager: Extract fingerprint logic
This commit extracts the fingerprint logic from the csi instance manager
into its own struct.

This change is to facilitate reusing the csimanager to also manage the
node-local CSI functionality, as it is the easiest place for us to
guaruntee health checking and to provide additional visibility into the
running operations through the fingerprinter mechanism and goroutine.
